requirements.txt
fastapi
uvicorn[standard]
python-multipart
 
PyMuPDF
pdfplumber
pdf2image
 
pytesseract
Pillow
numpy
pydantic
langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
rank_bm25